Fix glue interpolation
Sometimes an expression can be ambiguous.
Just because it parses, it does not mean that it
is an un-styled expression. So we also try to eval it.

Actually, even after eval-ing, they can be ambiguous, e.g.
{file /foo/bar} is probably a styled expression, but if
variables file, foo and bar exist, it is also a legal
R expression.